fix digest count (#2674)
* fix digest count
* fix digest count
* fix digest count
* Merge branch 'main' of https://git0.harness.io/l7B_kbSEQD2wjrM7PShm5w/PROD/Harness_Commons/gitness into AH-293-fix
* fix digest count

+++ registry/app/api/openapi/contracts/artifact/types.gen.go
@@ -106,7 +106,7 @@ type ArtifactSummary struct {
 
 // ArtifactVersionMetadata Artifact Version Metadata
 type ArtifactVersionMetadata struct {
-	DigestCount     *int64  `json:"digestCount,omitempty"`
+	DigestCount     *int    `json:"digestCount,omitempty"`
 	DownloadsCount  *int64  `json:"downloadsCount,omitempty"`
 	IslatestVersion *bool   `json:"islatestVersion,omitempty"`
 	LastModified    *string `json:"lastModified,omitempty"`
